requirements AI in defence.txt
Pillow==8.4.0
face-recognition==1.3.0
opencv-python-headless==4.5.4.58
numpy==1.21.4
